% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTaxa.R
\name{aggregateTaxa}
\alias{aggregateTaxa}
\title{Aggregate a taxon abundance matrix}
\usage{
aggregateTaxa(
  x,
  lineages,
  taxon.level = "class",
  unknown = c(NA, "unclassified"),
  keepUnknown = FALSE,
  keepSumUnknown = FALSE,
  returnLineages = FALSE
)
}
\arguments{
\item{x}{an abundance matrix with taxa as rows and samples as columns}

\item{lineages}{a lineage matrix where row names match row names of taxa; taxa must be provided in the same order as in x; first column gives kingdom, following columns give phylum, class, order, family, genus and species}

\item{taxon.level}{taxonomic level to which data set is aggregated (kingdom, phylum, class, order, family, genus or species)}

\item{unknown}{descriptors for unclassified taxa}

\item{keepUnknown}{keep taxa without higher-level taxon classification; cannot be used together with keepSumUnknown}

\item{keepSumUnknown}{keep the sum of taxa without higher-level taxon classification on the desired level as last row (row name: sumUnknown)}

\item{returnLineages}{return lineage matrix for higher-level taxon matrix; if true a list is returned with two elements: abundances and lineages}
}
\value{
the aggregated abundance matrix; if returnLineages is true, a list with aggregated abundances and matching lineages
}
\description{
Given a taxon and a lineage matrix, aggregate taxa belonging to the same
higher-level taxon at the specified taxonomic level by summing their abundance.
}
\examples{
data("ibd_taxa")
data("ibd_lineages")
ibd_class=aggregateTaxa(x=ibd_taxa,lineages=ibd_lineages)
}
