% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMetadata.R
\name{assignMetadataTypes}
\alias{assignMetadataTypes}
\title{Assign data types to metadata}
\usage{
assignMetadataTypes(data, categoric = c())
}
\arguments{
\item{data}{a metadata matrix with rows as samples and columns as items or a dataframe}

\item{categoric}{metadata items to be treated as categoric data}
}
\description{
Treat all metadata with more than two values
as numeric, unless they are in the vector with categoric values.
Metadata with two values (without counting missing values) are treated as
binary (so categoric).
Metadata are supposed to have been filtered previously (so constant rows were removed).
Note that a message is printed to indicate which metadata type was assigned and to
warn about numeric metadata with non-numeric values.
}
