% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Build a microbial network}
\usage{
buildNetwork(
  abundances,
  lineages = NULL,
  min.occ = 0,
  norm = FALSE,
  clr = FALSE,
  method = "mb",
  repNum = 20,
  minStrength = 0.5,
  alpha = 0.05,
  renorm = FALSE,
  initEdgeNum = 100,
  methodNum = 2,
  directed = FALSE,
  nameLevel = "Genus",
  colorLevel = "Class",
  widthFactor = 10,
  legend = TRUE,
  left.margin = 10
)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{lineages}{a matrix with lineages where row names match row names of abundances; first column gives kingdom, following columns give phylum, class, order, family, genus and species (obligatory for SPIEC-EASI)}

\item{min.occ}{only keep rows with at least the given number of non-zero values, keep the sum of the other rows (carried out before network construction)}

\item{norm}{normalize matrix (carrried out after filtering); not carried out for SPIEC-EASI}

\item{clr}{apply CLR transform (after filtering and normalization; \code{\link{clr}} with omit.zeros true); not carried out for SPIEC-EASI}

\item{method}{mb or glasso (SPIEC-EASI), any combination of pearson, spearman, bray and/or kld (CoNet), hc, tabu, h2pc, mmhc, rsmax2, aracne, pc.stable, gs, fast.iamb, iamb.fdr, mmpc, hpc or si.hiton.pc (bnlearn, for aracne mi is set to mi-g)}

\item{repNum}{the number of permutation/bootstrap iterations for CoNet with dissimilarities and of bootstrap iterations in SPIEC-EASI and bnlearn; can be omitted (set to 0) for time-intensive bnlearn methods or for CoNet without p-values}

\item{minStrength}{bnlearn only: the minimum probability for an arc to appear across bootstraps (only applicable if repNum is set larger 1)}

\item{alpha}{p-value threshold for bnlearn and CoNet (ignored in methods aracne, hc, tabu, rsmax2, mmhc and h2pc)}

\item{renorm}{compute correlation p-values in CoNet with renorm enabled (slow)}

\item{initEdgeNum}{CoNet parameter for the number of initial top and bottom edges}

\item{methodNum}{CoNet threshold on the minimum method number needed to keep an edge (only used when more than one method is provided)}

\item{directed}{bnlearn only: infer a directed network (hc, tabu, rsmax2, mmhc and h2pc only infer directed networks, aracne only undirected networks)}

\item{nameLevel}{the taxonomic level used for node names (only if lineages provided)}

\item{colorLevel}{the taxonomic level used for node colors (only if lineages provided)}

\item{widthFactor}{edge width scaling factor}

\item{legend}{if lineages are provided: show node color code}

\item{left.margin}{if lineages are provided and legend is true: set the left margin (the margin is restored to its previous value afterwards)}
}
\value{
igraph object
}
\description{
Wrapper for microbial network inference methods.
SPIEC-EASI, CoNet (see function \code{\link{barebonesCoNet}}) and bnlearn are supported.
Abundances are assumed to have been already filtered and transformed.
The function returns the network as an igraph object,
with nodes labeled and colored at the selected taxonomic levels
and, for SPIEC-EASI and CoNet, edges in red or green, depending on whether they
represent negative or positive associations.
}
\details{
For SPIEC-EASI, edge width represents the abolute of the regression coefficients, whereas for bnlearn,
it represents the probability of an arc to re-appear across bootstrap iterations. For CoNet, it either
represents absolute association strength, method number supporting an edge or significance,
depending on whether more than one method was used to compute the network and whether or not
p-values were computed (significance is defined as -log10(pval)). For single dissimilarities in CoNet, edge weight
is scaled (see \code{\link{barebonesCoNet}} for details) such that for all network construction methods,
a thicker edge means a stronger/more stable/more significant association.
For CoNet, p-values are computed by default with cor.test for correlations and permutations and bootstraps
for dissimilarities and are multiple-testing corrected with Benjamini-Hochberg.
P-value computation in CoNet can be disabled by setting repNum to 0.
If lineages are provided and legend is true, the color code for the node colors is plotted in a separate bar plot.
SPIEC-EASI and CoNet return undirected graphs, whereas bnlearn methods support the inference
of directed graphs. SPIEC-EASI methods tend to be faster than bnlearn methods.
The returned network can be visualized with igraph's interactive network visualisation
function tkplot or simply with igraph's plot function. When RCy3 is installed and Cytoscape is open,
the graph can also be sent to Cytoscape directly with RCy3's function createNetworkFromIgraph().
}
\examples{
data(ibd_taxa)
data(ibd_lineages)
top.indices=sort(rowSums(ibd_taxa),decreasing = TRUE,index.return=TRUE)$ix[1:30]
taxa=ibd_taxa[top.indices,]
lineages=ibd_lineages[top.indices,]
hc.out=buildNetwork(taxa,lineages,method="hc",nameLevel="species")
plot(hc.out)
}
