% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barebonesCoNet.R
\name{clr}
\alias{clr}
\title{Centered log-ratio (CLR) transform}
\usage{
clr(
  abundances,
  minocc = 0,
  pseudocount = 1,
  scale.factor = 1,
  omit.zeros = FALSE
)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{minocc}{prevalence filter; ignored if 0 or below}

\item{pseudocount}{zero replacement value; should be above 0}

\item{scale.factor}{scaling factor; abundances will be multiplied with this factor}

\item{omit.zeros}{if TRUE, ignore pseudocount and compute clr transform only on non-zero elements}
}
\value{
clr-transformed matrix
}
\description{
In the CLR transform, each entry is divided by the geometric mean of its sample
and then the logarithm is taken.
}
\details{
When omit.zeros is set to true, the default behaviour of clr in the compositions package
when applied column-wise is reproduced.
}
