% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGroups.R
\name{compareGroups}
\alias{compareGroups}
\title{Compare diversity-related properties across groups}
\usage{
compareGroups(
  abundances,
  property = "beta",
  method = "dissim",
  groups = c(),
  colors = c(),
  plot.type = "pergroup",
  avg = "none",
  all = FALSE,
  noSameGroup = TRUE,
  rowNorm = FALSE,
  subsample = TRUE,
  xlab = "",
  pvalViz = FALSE
)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{property}{richness (Chao1), evenness (Sheldon), alpha (alpha-diversity with Shannon index) or beta (beta-diversity)}

\item{method}{for beta-diversity, dissim (Bray Curtis dissimilarities) and DM (estimation of theta with package dirmult) are supported}

\item{groups}{group membership vector with as many entries as samples in abundances}

\item{colors}{define a color for each group with as many entries as there are groups, appearing in the order in which groups first appear (only for plot.type pergroup)}

\item{plot.type}{pergroup (plot property per group; this is the only plot type available for richness, evenness, alpha and beta with DM) and intravsinter (plot property for all within-group samples versus all samples)}

\item{avg}{none, mean or median (average group property with the selected averaging method, does not work for method DM)}

\item{all}{include the beta diversity for all samples}

\item{noSameGroup}{for beta diversity with method dissim: only between-group pairs are used to calculate the beta diversity for all samples}

\item{rowNorm}{for beta diversity with method dissim: normalize the abundances row-wise}

\item{subsample}{subsample groups randomly to have the same sample number in each group; for plot type intravsinter, it equalises in additon the number of pairs within and between groups}

\item{xlab}{the x axis label}

\item{pvalViz}{if true and avg is set to none and plot.type is pergroup, significant Wilcoxon p-values are displayed on the box plot using function stat_compare_means in R package ggpubr}
}
\description{
The function allows to compare richness, evenness, alpha diversity and beta diversity across groups, where
beta diversity can be assessed with both a Bray Curtis distribution or overdispersion from a Dirichlet-Multinomial distribution fit.
By default, groups are subsampled to the same sample number, for this reason the output of the same run can differ. To switch off this
behaviour, subsample can be set to false. Chao1 and Shannon diversity are computed with vegan.
}
\examples{
data("ibd_taxa")
data("ibd_metadata")
groups=as.vector(ibd_metadata$Diagnosis)
compareGroups(ibd_taxa,groups=groups,property="alpha",pvalViz = TRUE)
}
