% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barebonesCoNet.R
\name{filterTaxonMatrix}
\alias{filterTaxonMatrix}
\title{Filter taxa in an abundance matrix}
\usage{
filterTaxonMatrix(
  x,
  minocc = 0,
  keepSum = FALSE,
  return.filtered.indices = FALSE
)
}
\arguments{
\item{x}{taxon abundance matrix, rows are taxa, columns are samples}

\item{minocc}{minimum occurrence (minimum number of samples with non-zero taxon abundance)}

\item{keepSum}{If keepSum is true, the discarded rows are summed and the sum is added as a row with name: summed-nonfeat-rows}

\item{return.filtered.indices}{if true, return an object with the filtered abundance matrix in mat and the indices of removed taxa in the original matrix in filtered.indices}
}
\value{
filtered abundance matrix
}
\description{
Discard taxa with less than the given minimum number of occurrences.
}
