% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findClusters.R
\name{findClusters}
\alias{findClusters}
\title{Cluster taxon- or sample-wise}
\usage{
findClusters(abundances, method = "dmm", k = 5)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{method}{clustering method, supported are dmm and pam}

\item{k}{cluster number}
}
\value{
a cluster membership vector
}
\description{
Cluster sequencing data taxon- or sample-wise.
By default, data are clustered sample-wise. The default method is
Dirichlet-Multinomial mixtures (DMM) using the DirichletMultinomial package.
Note that DMM expects counts and fails if there are taxa that are absent across all samples.
Also note that counts should not be too large for DMM (not above 10000). If this is the case,
they will be scaled (divided by a constant and rounded). In addition, for DMM, the total number
of counts should be the same across samples.
Partitioning around medoids (PAM) does not expect counts, can deal with absent taxa and large counts.
}
