% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floresgut_lineages.R
\docType{data}
\name{floresgut_lineages}
\alias{floresgut_lineages}
\title{Lineages for gut taxa from Flores et al.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 323 rows and 7 columns.
}
\usage{
data(floresgut_lineages)
}
\value{
Taxon x phylogenetic level matrix of size 323 x 7
}
\description{
Lineages include the levels kingdom, phylum, class, order, family, genus and species.
}
\details{
Row names match taxon names in floresgut_taxa.
}
\examples{
data(floresgut_lineages)
}
\references{
Flores et al. (2014) Temporal variability is a personalized feature of the human microbiome Genome Biology 15, 531
\href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0531-y}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
