% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floresgut_metadata.R
\docType{data}
\name{floresgut_metadata}
\alias{floresgut_metadata}
\title{Gut sample metadata from Flores et al.}
\format{
An object of class \code{data.frame} with 637 rows and 11 columns.
}
\usage{
data(floresgut_metadata)
}
\value{
Sample x metadata data frame of size 637 x 11
}
\description{
Metadata are selected from the sample information file in QIITA (ID: 2151) and include age, bmi, diet, gender,
host_subject_id, ibd, menstruation_disturbance, raceethnicity, sickness_disturbance, smokecigarettes and weeknumber.
Metadata is a data frame with samples as rows and metadata items as columns. Samples are in the same order as the samples in
floresgut_taxa.
}
\details{
Note that some metadata items contain missing values.
}
\examples{
data(floresgut_metadata)
}
\references{
Flores et al. (2014) Temporal variability is a personalized feature of the human microbiome Genome Biology 15, 531
\href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0531-y}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
