% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floresgut_taxa.R
\docType{data}
\name{floresgut_taxa}
\alias{floresgut_taxa}
\title{Gut taxon abundances from Flores et al.}
\format{
An object of class \code{matrix} with 323 rows and 637 columns.
}
\usage{
data(floresgut_taxa)
}
\value{
Taxon x sample matrix of size 323 x 637
}
\description{
Data were obtained through 16S sequencing of the V4 region.
}
\details{
OTU abundances were downloaded from QIITA in Nov 2020 (ID: 2151). Control and non-gut samples, samples with less than 10000 reads,
mislabel chance above 50% or no mislabel data were removed. Samples belonging to subjects with less than 6 samples taken
were also removed. Samples were rarefied to 10000 reads and OTUs were aggregated to genus level. OTUs without a known genus were summed into
a garbage taxon named sumUnknown. The resulting genus count matrix has 323 taxa and 637 samples.
}
\examples{
data(floresgut_taxa)
}
\references{
Flores et al. (2014) Temporal variability is a personalized feature of the human microbiome Genome Biology 15, 531
\href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-014-0531-y}{Genome Biology}
}
\keyword{data,}
\keyword{datasets}
