% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barebonesCoNet.R
\name{get.kld}
\alias{get.kld}
\title{Compute KLD between two vectors}
\usage{
get.kld(x, y, pseudocount = 1e-08)
}
\arguments{
\item{x}{vector with non-negative numbers}

\item{y}{vector with non-negative numbers}

\item{pseudocount}{pseudocount = this value is added to each zero}
}
\value{
kullback-leibler dissimilarity value
}
\description{
Compute Kullback-Leibler dissimilarity
}
\details{
Outputs a Kullback-Leibler dissimilarity (symmetrized divergence) matrix. Note:
}
