% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupBarplot.R
\name{groupBarplot}
\alias{groupBarplot}
\title{Bar plot of taxon composition with group support}
\usage{
groupBarplot(abundances, groups = c(), taxon.color.map = NULL,
  group.color.map = NULL, topTaxa = 10, sortGroupwise = TRUE,
  randSampleNum = NA, summedTaxonColor = "#a9a9a9",
  extendTaxonColorMap = FALSE, legend = TRUE, legend.shift = 0.25,
  ...)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{groups}{group membership vector with as many entries as samples}

\item{taxon.color.map}{map of taxon-specific colors, should match row names; taxa not present in the color map will be colored in summedTaxonColor}

\item{group.color.map}{map of group-specific colors, should match group names}

\item{topTaxa}{number of top taxa to be plotted}

\item{sortGroupwise}{if true, samples are sorted according to groups}

\item{randSampleNum}{if larger 0, sortGroupwise is set to true and the indicated sample number is randomly selected for each group}

\item{summedTaxonColor}{the color of the summed taxa, by default gray}

\item{extendTaxonColorMap}{if true, taxa not in the taxon color map are added there and the extended color map is returned}

\item{legend}{add a legend with the color code}

\item{legend.shift}{decrease this parameter to shift the legend further to the right}

\item{\\dots}{Additional arguments passed to plot()}
}
\description{
Plot taxa in a bar plot sorted by summed abundance across samples
}
\examples{
data(ibd_taxa)
data(ibd_metadata)
# taxon abundances were prefiltered and therefore do not add up to 1
groupBarplot(ibd_taxa,groups=as.vector(ibd_metadata$Diagnosis),randSampleNum=7)
data(ibd_lineages)
ibd_genera=aggregateTaxa(ibd_taxa,ibd_lineages,taxon.level = "genus")
groupBarplot(ibd_genera,groups=as.vector(ibd_metadata$Diagnosis),randSampleNum=7)
}
