% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_lineages.R
\docType{data}
\name{ibd_lineages}
\alias{ibd_lineages}
\title{Lineages for taxa from the PRISM cohort}
\format{An object of class \code{matrix} with 201 rows and 7 columns.}
\usage{
data(ibd_lineages)
}
\value{
Species x phylogenetic level matrix of size 201 x 7
}
\description{
Lineages include the levels kingdom, phylum, class, order, family, genus and species.
}
\details{
Row names match taxon names in ibd_taxa.
}
\examples{
data(ibd_lineages)
}
\references{
Franzosa et al. (2019) Gut microbiome structure and metabolic activity in inflammatory bowel disease Nature Microbiology 4, 293-305
\href{https://www.nature.com/articles/s41564-018-0306-4}{Nature Microbiology}
}
\keyword{data,}
\keyword{datasets}
