% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_metadata.R
\docType{data}
\name{ibd_metadata}
\alias{ibd_metadata}
\title{Metadata for the PRISM cohort}
\format{An object of class \code{data.frame} with 155 rows and 8 columns.}
\usage{
data(ibd_metadata)
}
\value{
Sample x metadata dataframe of size 155 x 8
}
\description{
Metadata include SRA identifier, age, diagnosis, fecal calprotectin, antibiotic and immunosuppressant usage as well as treatment with mesalamine or steroids.
}
\details{
Stool samples were taken from patients with Crohn's disease (CD, n=68) and ulcerative colitis (UC, n=53) as well as healthy controls (n=34).
CD/UC vs healthy status is stored in 'Diagnosis'. The order of samples matches those of ibd_taxa.
}
\examples{
data(ibd_metadata)
# convert to numeric
ibd_metadata$Fecal.Calprotectin=as.numeric(ibd_metadata$Fecal.Calprotectin)
ibd_metadata$Age=as.numeric(ibd_metadata$Age)
summary(ibd_metadata)
}
\references{
Franzosa et al. (2019) Gut microbiome structure and metabolic activity in inflammatory bowel disease Nature Microbiology 4, 293-305
\href{https://www.nature.com/articles/s41564-018-0306-4}{Nature Microbiology}
}
\keyword{data,}
\keyword{datasets}
