% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibd_taxa.R
\docType{data}
\name{ibd_taxa}
\alias{ibd_taxa}
\title{IBD and control taxon abundances from the PRISM cohort}
\format{An object of class \code{matrix} with 201 rows and 155 columns.}
\usage{
data(ibd_taxa)
}
\value{
Species x sample matrix of size 201 x 155
}
\description{
Species abundances extracted from whole-genome shotgun data of stool samples.
}
\details{
Samples include patients with Crohn's disease (CD, n=68) and ulcerative colitis (UC, n=53) as well as healthy controls (n=34).
Species relative abundances were computed with MetaPhlAn2 and species below or equal to 0.1% abundance in at least five samples were excluded.
}
\examples{
data(ibd_taxa)
}
\references{
Franzosa et al. (2019) Gut microbiome structure and metabolic activity in inflammatory bowel disease Nature Microbiology 4, 293-305
\href{https://www.nature.com/articles/s41564-018-0306-4}{Nature Microbiology}
}
\keyword{data,}
\keyword{datasets}
