% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTables.R
\name{intersectTables}
\alias{intersectTables}
\title{Table intersection}
\usage{
intersectTables(table1, table2, byRow = FALSE)
}
\arguments{
\item{table1}{a matrix}

\item{table2}{a matrix}

\item{byRow}{compare tables row-wise}
}
\value{
a list with the two tables with matching sample names
}
\description{
Discard rows or columns that are not present
in both tables (columns by default). The row or column
names are compared to determine whether they match.
}
