% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeATC.R
\name{mergeATC}
\alias{mergeATC}
\title{Merge ATC codes}
\usage{
mergeATC(data, level = 1)
}
\arguments{
\item{data}{matrix with only zeros or ones, with ATC codes as rows and samples as columns}

\item{level}{ATC level to which data should be merged}
}
\value{
matrix with merged ATC codes
}
\description{
ATC codes are a classification system for therapeutic chemicals, see
https://en.wikipedia.org/wiki/Anatomical_Therapeutic_Chemical_Classification_System
}
