% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTables.R
\name{mergeTables}
\alias{mergeTables}
\title{Table union}
\usage{
mergeTables(table1, table2, keep.nonmatch = FALSE, byRow = TRUE)
}
\arguments{
\item{table1}{a matrix}

\item{table2}{a matrix}

\item{keep.nonmatch}{keep non-matching rows or columns, with values missing in the other table filled with zeros}

\item{byRow}{compare tables row-wise}
}
\value{
the merged table
}
\description{
Merge two matrices row- or column-wise using their row
or column names (row names by default). Expects two matrices and appends
columns of the second table to the first such that row names in the
second table match row names in the first.
Rows with non-matching names are discarded by default, but can
be kept. In this case, their values are set to zero
in the table from which they are absent.
}
