% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkIntersectSig.R
\name{networkIntersectSig}
\alias{networkIntersectSig}
\title{Significance of network intersection}
\usage{
networkIntersectSig(
  rep.num = 5,
  taxon.num = 20,
  avg.network.size = 0,
  inter.size = 0,
  iter = 1000,
  directed = FALSE,
  loops = FALSE,
  hist.rand = FALSE
)
}
\arguments{
\item{rep.num}{number of inferred networks}

\item{taxon.num}{number of taxa in the matrix}

\item{avg.network.size}{average number of edges in inferred networks}

\item{inter.size}{number of edges in observed intersection network}

\item{iter}{number of iterations for random network group construction}

\item{directed}{if directed, the possible edge number is taxon.num * (taxon.num-1), else it is (taxon.num * (taxon.num-1))/2}

\item{loops}{if directed, count self-arcs (so compute possible edge number as taxon.num * taxon.num)}

\item{hist.rand}{if TRUE, plot the histogram of random intersection sizes}
}
\value{
p-value
}
\description{
Builds groups of networks randomly and counts their number of intersection edges (intersection size).
Each random network is built by randomly sampling from the total number of possible edges.
The p-value is computed parameter-free by counting the number of times the intersection size of
random networks is smaller than the intersection size in the observed networks and dividing it
by the number of iterations.
}
\examples{
networkIntersectSig(rep.num=3,taxon.num=8,avg.network.size=12,inter.size=10,directed=TRUE)
}
