% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqPCoA.R
\name{rarefyFilter}
\alias{rarefyFilter}
\title{Rarefaction combined with sample filtering}
\usage{
rarefyFilter(x, min = 0)
}
\arguments{
\item{x}{a matrix}

\item{min}{minimum count to which x is to be rarefied (if equal to zero, the minimum column sum is taken as min)}
}
\value{
a list with the rarefied matrix (rar) and the indices of the columns that were kept (colindices)
}
\description{
Rarefy a matrix to the given minimum count number column-wise
using vegan's rrarefy function. If columns have less than the minimum count number,
they are discarded. Rows that have a sum of zero after rarefaction are also discarded.
}
