% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMetadata.R
\name{removeGroupsWithMissingValues}
\alias{removeGroupsWithMissingValues}
\title{Remove groups that have metadata items with missing values}
\usage{
removeGroupsWithMissingValues(data, groups = c())
}
\arguments{
\item{data}{a metadata matrix with rows as samples and columns as items or a dataframe with metadata items as columns}

\item{groups}{group membership vector with as many entries as samples}
}
\value{
a vector with indices of samples without problematic groups
}
\description{
Remove groups that have metadata items with missing values
}
