% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSamples.R
\name{selectGroups}
\alias{selectGroups}
\title{Select groups}
\usage{
selectGroups(
  groups = c(),
  selected.groups = c(),
  inverse = FALSE,
  returnIndex = FALSE
)
}
\arguments{
\item{groups}{membership vector with as many entries as samples}

\item{selected.groups}{names of groups to keep}

\item{inverse}{keep groups that are not in selected groups}

\item{returnIndex}{return the indices of the selected samples}
}
\value{
modified group membership vector
}
\description{
Select groups
}
