% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqgroup.R
\docType{package}
\name{seqgroup}
\alias{seqgroup}
\title{Analysis of Sequencing Data with Groups}
\description{
The seqgroup R package provides functions that support the analysis of metadata and groups in microbial sequencing data.
}
