% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqPCoA.R
\name{seqPCoA}
\alias{seqPCoA}
\title{PCoA for microbial sequencing data}
\usage{
seqPCoA(abundances, metadata = NULL, groupAttrib = "", groups = c(),
  clusters = c(), labels = c(), time = c(), hiddenSamples = c(),
  dis = "bray", rda = FALSE, scale = FALSE, doScree = FALSE,
  topTaxa = 10, topMetadata = 10, arrowFactor = 0.5,
  metadataFactor = 1, centroidFactor = 1, taxonColor = "brown",
  metadataColor = "blue", xlim = c(-0.3, 0.3), ylim = c(-0.3, 0.3),
  permut = 1000, pAdjMethod = "BH", qvalThreshold = 0.05,
  dimensions = c(1, 2), ...)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{metadata}{an optional data frame with metadata items, where samples are in the same order as in x, if provided and rda is FALSE, envfit is carried out}

\item{groupAttrib}{optional: the name of a metadata item that refers to a vector that provides for each sample its group membership}

\item{groups}{an optional vector that provides for each sample its group membership and which is overridden by groupAttrib, if provided}

\item{clusters}{an optional vector that provides for each sample its cluster membership as an integer (cluster membership is visualized through shape, up to 10 different shapes are possible)}

\item{labels}{an optional vector that provides for each sample its label to display}

\item{time}{an optional vector with as many time points as samples, adds arrows between consecutive time points}

\item{hiddenSamples}{an optional vector with indices of samples to be hidden (they will be taken into account for PCoA/RDA/envfit, but are not displayed)}

\item{dis}{dissimilarity or distance supported by vegan}

\item{rda}{carry out an RDA instead of a PCoA}

\item{scale}{scale numeric metadata (subtract the mean and divide by standard deviation)}

\item{doScree}{do a Scree plot}

\item{topTaxa}{if larger than zero: show the top N taxa most correlated to principal components as arrows in the PCoA}

\item{topMetadata}{if larger than zero, metadata provided and rda false: show the top N most significant numeric metadata as arrows and the top N most significant factor metadata as text in the PCoA}

\item{arrowFactor}{the length of taxon arrows (determined by scaled covariance) is multiplied with this factor}

\item{metadataFactor}{the length of numeric metadata arrows (determined by Pearson correlation) is multiplied with this factor}

\item{centroidFactor}{centroid positions are multiplied with this factor}

\item{taxonColor}{the color of the taxon arrows and text}

\item{metadataColor}{the color of the metadata arrows and text}

\item{xlim}{range shown on the x axis}

\item{ylim}{range shown on the y axis}

\item{permut}{number of permutations in envfit}

\item{pAdjMethod}{method for multiple testing correction supported by p.adjust for envfit p-values}

\item{qvalThreshold}{threshold on multiple-testing corrected envfit p-values}

\item{dimensions}{the principal components used for plotting, by default the first and second}

\item{\\dots}{Additional arguments passed to plot()}
}
\description{
A wrapper around various PCoA-based analyses implemented in vegan. The wrapper can handle groups and
metadata. The na.action is set to na.omit, however envfit cannot deal with
missing values, therefore if metadata are provided, they should be free of missing values.
}
