% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMetadata.R
\name{setNAToGroupMean}
\alias{setNAToGroupMean}
\title{Replace missing values by group mean.}
\usage{
setNAToGroupMean(metadata.df, groups = c(), na.threshold = 4,
  metadata.to.skip = c())
}
\arguments{
\item{metadata.df}{a data frame with metadata}

\item{groups}{a vector that specifies the group for each sample}

\item{na.threshold}{number of allowed missing values per group}

\item{metadata.to.skip}{metadata for which filling with the group mean does not make sense (e.g. date)}
}
\description{
Missing values are replaced either by the group mean
for numeric metadata or the most frequent group value for categoric
metadata. Missing values are only replaced if there are less
missing values than the threshold in the group, else the metadata item
concerned is removed. The output are metadata without missing values.
}
