% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGroups.R
\name{sheldon}
\alias{sheldon}
\title{It ranges from 0 to 1, where 1 signifies a perfectly even abundance distribution.}
\usage{
sheldon(x, correction = TRUE, N2N1 = FALSE)
}
\description{
It ranges from 0 to 1, where 1 signifies a perfectly even abundance distribution.
}
