% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSamples.R
\name{sortSamples}
\alias{sortSamples}
\title{Sort samples by group and optionally by time.}
\usage{
sortSamples(abundances, groups = c(), time = c(), alphabetical = FALSE)
}
\arguments{
\item{abundances}{an abundance matrix with taxa as rows and samples as columns}

\item{groups}{group membership vector with as many entries as samples}

\item{time}{optional time vector with as many entries as samples; entries are numeric or characters of numbers}

\item{alphabetical}{option to order samples by group name in alphabetical order}
}
\value{
an index vector with the updated sample order
}
\description{
Sort samples by group and optionally by time.
}
