% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toPhyloseq.R
\name{toPhyloseq}
\alias{toPhyloseq}
\title{Convert abundances to phyloseq object}
\usage{
toPhyloseq(
  abundances,
  metadata = NULL,
  lineages = NULL,
  assignSampleNames = FALSE,
  dummyMetadata = FALSE,
  dummyLineages = FALSE,
  lineageSource = NULL
)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{metadata}{an optional data frame with metadata items as columns, where samples are in the same order as in abundances}

\item{lineages}{an optional lineage matrix where row names match row names of taxa; taxa must be provided in the same order as in x; first column gives kingdom, following columns give phylum, class, order, family, genus and species}

\item{assignSampleNames}{assign random sample names}

\item{dummyMetadata}{create a dummy metadata object with a single numeric item from a normal distribution; only if no metadata are provided}

\item{dummyLineages}{assign dummy lineages; only if no lineages are provided and a lineageSource is given}

\item{lineageSource}{a phyloseq object with known lineages, e.g. GlobalPatterns (phyloseq preloaded data that can be loaded with data(GlobalPatterns)), to be used for random lineage assignment}
}
\value{
phyloseq object
}
\description{
Convert abundances to phyloseq object
}
\examples{
data(ibd_taxa)
data(ibd_metadata)
data(ibd_lineages)
phyloseq.obj=toPhyloseq(ibd_taxa,metadata=ibd_metadata,lineages=ibd_lineages)
phyloseq.obj
}
