% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateTaxa.R
\name{aggregateTaxa}
\alias{aggregateTaxa}
\title{Aggregate a taxon abundance matrix}
\usage{
aggregateTaxa(x, lineages, taxon.level = "class", unknown = c(NA,
  "unclassified"))
}
\arguments{
\item{x}{an abundance matrix with taxa as rows and samples as columns}

\item{lineages}{a lineage matrix where row names match row names of taxa; taxa must be provided in the same order as in x; first column gives kingdom, following columns give phylum, class, order, family, genus and species}

\item{taxon.level}{taxonomic level to which data set is aggregated (kingdom, phylum, class, order, family, genus or species)}

\item{unknown}{descriptors of unclassified taxa}
}
\description{
Given a taxon and a lineage matrix, aggregate taxa belonging to the same
higher-level taxon at the specified taxonomic level by summing their abundance.
}
