% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqPCoA.R
\name{assignColorsToGroups}
\alias{assignColorsToGroups}
\title{Assign colors to groups}
\usage{
assignColorsToGroups(
  groups,
  refName = "ref",
  myColors = NULL,
  returnMap = FALSE,
  hiddenSamples = c()
)
}
\arguments{
\item{groups}{a vector of group memberships}

\item{refName}{the name of the reference group; receives gray as a color}

\item{myColors}{an optional map of predefined colors for groups}

\item{returnMap}{whether to return the color map together with the colors}

\item{hiddenSamples}{optional vector of indices of samples to be hidden; these are assigned an invisible color}
}
\value{
a vector of colors; if returnMap is true, a vector of colors and the color map
}
\description{
Given a group membership vector, each group receives
a unique color such that the resulting color vector is as long as
the group membership vector.
}
