% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{buildNetwork}
\alias{buildNetwork}
\title{Build a microbial network}
\usage{
buildNetwork(abundances, lineages, method = "mb", repNum = 20,
  nameLevel = "Genus", colorLevel = "Class", widthFactor = 50)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{lineages}{a matrix with lineages in the phyloseq format}

\item{method}{mb and glasso, supported by SPIEC-EASI}

\item{repNum}{the number of bootstrap iterations in SPIEC-EASI}

\item{nameLevel}{the taxonomic level used for node names}

\item{colorLevel}{the taxonomic level used for node colors}

\item{widthFactor}{edge width scaling factor (for SPIEC-EASI, the width is the abolute of the regression coefficients)}
}
\value{
the SPIEC-EASI igraph object
}
\description{
Wrapper for microbial network inference methods.
For the moment, only SPIEC-EASI is supported.
Abundances are assumed to have been already filtered.
The function returns the network as an igraph object,
with nodes labeled and colored at the selected taxonomic levels
and edges in red or green, depending on whether they
represent negative or positive associations. Edge thickness
represents association strength. The color code for the node colors is
plotted in a separate bar plot.
The returned network can be visualized with igraph's
interactive network visualisation function tkplot or
simply with igraph's plot function.
}
