% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDistribs.R
\name{compareDistribs}
\alias{compareDistribs}
\title{Compare the distribution of a taxon across two groups}
\usage{
compareDistribs(
  x,
  taxon,
  groups = c(),
  group1 = 1,
  group2 = 2,
  name1 = as.character(group1),
  name2 = as.character(group2)
)
}
\arguments{
\item{x}{a matrix with taxa as rows and samples as columns}

\item{taxon}{the index or name of a taxon to be compared}

\item{groups}{a vector that provides for each sample its group membership (numeric or character)}

\item{group1}{the identifier of the first group (numeric or character)}

\item{group2}{the identifier of the second group (numeric or character)}

\item{name1}{the name to be displayed for the first group}

\item{name2}{the name to be displayed for the second group}
}
\description{
Plot taxon abundances as histograms and calculate the
unpaired two-sided Wilcoxon p-value.
}
\examples{
data("ibd_taxa")
data("ibd_metadata")
groups=as.vector(ibd_metadata$Diagnosis)
taxon="Faecalibacterium_prausnitzii"
compareDistribs(ibd_taxa,taxon=taxon,groups=groups,group1="UC",group2="Control")
groups[groups=="UC"]="IBD"
groups[groups=="CD"]="IBD"
# Faecalibacterium is significantly more abundant in the control group
compareDistribs(ibd_taxa,taxon=taxon,groups=groups,group1="IBD",group2="Control")
}
