% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareGroups.R
\name{compareGroups}
\alias{compareGroups}
\title{Compare diversity-related properties across groups}
\usage{
compareGroups(abundances, property = "beta", method = "dissim",
  groups = c(), plot.type = "pergroup", avg = "none", all = FALSE,
  noSameGroup = TRUE, rowNorm = FALSE, subsample = TRUE,
  xlab = "Groups", pvalViz = FALSE)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{property}{richness (Chao1), evenness (Sheldon), alpha (alpha-diversity with Shannon index) or beta (beta-diversity)}

\item{method}{for beta-diversity, dissim (Bray Curtis dissimilarities) and DM (estimation of theta with the Dirichlet-Multnomial) are supported}

\item{groups}{group membership vector with as many entries as samples in abundances}

\item{plot.type}{pergroup (plot property per group; this is the only plot type available for richness, evenness, alpha and beta with DM) and intravsinter (plot property for all within-group samples versus all samples)}

\item{avg}{none, mean or median (average group property with the selected averaging method, does not work for method DM)}

\item{all}{include the beta diversity for all samples}

\item{noSameGroup}{for beta diversity with method dissim: only between-group pairs are used to calculate the beta diversity for all samples}

\item{rowNorm}{for beta diversity with method dissim: normalize the abundances row-wise}

\item{subsample}{for intravsinter plot type: subsample to have as many intra-group as inter-group pairs prior to performing a Wilcoxon test}

\item{xlab}{the x axis label}

\item{pvalViz}{if true and avg is set to none and plot.type is pergroup, significant Wilcoxon p-values are displayed on the box plot}
}
\description{
Compare diversity-related properties across groups
}
