% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterMetadata.R
\name{filterMetadata}
\alias{filterMetadata}
\title{Filter metadata}
\usage{
filterMetadata(data, toFilter = c(), na.threshold = 0)
}
\arguments{
\item{data}{a matrix where metadata are rows and samples are columns or a data frame of metadata items}

\item{toFilter}{names of metadata to remove}

\item{na.threshold}{percentage of NA values allowed, between 0 and 100}
}
\value{
a data frame of filtered metadata items
}
\description{
Remove metadata with more than the given percentage
of missing values, constant metadata and metadata with names as provided
in the toFilter vector.
}
