% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoreTaxa.R
\name{getCoreTaxa}
\alias{getCoreTaxa}
\title{List core taxa across groups}
\usage{
getCoreTaxa(x, groups = c(), min.occ = 1)
}
\arguments{
\item{x}{an abundance matrix where rows are taxa and columns samples}

\item{groups}{group membership vector with as many entries as abundances has samples}

\item{min.occ}{minimal occurrence to count a taxon as present in each group}
}
\value{
indices of core taxa
}
\description{
List the indices of all taxa that occur at least min.occ times in each group.
}
\examples{
abundances=cbind(c(1,2,2),c(0,1,1),c(0,0,3))
rownames(abundances)=c("taxon1","taxon2","taxon3")
rownames(abundances)[getCoreTaxa(abundances,groups=c(1,2,3))]
}
