% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqPCoA.R
\name{seqPCoA}
\alias{seqPCoA}
\title{PCoA for microbial sequencing data}
\usage{
seqPCoA(abundances, reference = NULL, rarefyRef = FALSE,
  refName = "ref", metadata = NULL, groupAttrib = "", groups = c(),
  groupColors = NULL, colors = c(), clusters = c(), labels = c(),
  sizes = c(), size.legend = "", time = c(), hiddenSamples = c(),
  dis = "bray", rda = FALSE, scale = FALSE, doScree = FALSE,
  topTaxa = 10, topMetadata = 10, arrowFactor = 0.5,
  metadataFactor = 1, centroidFactor = 1, taxonColor = "brown",
  metadataColor = "blue", xlim = NULL, ylim = NULL, permut = 1000,
  env.permut = 1000, pAdjMethod = "BH", qvalThreshold = 0.05,
  dimensions = c(1, 2), ...)
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{reference}{an optional reference data set on which abundances are mapped; data are merged by matching row names (nonmatching ones are kept as sum); cannot be combined with rda or topMetadata (topMetadata needs to be set to zero)}

\item{rarefyRef}{rarefy abundance and reference samples to the minimum total count found in any of the samples; recommended when the total counts differ}

\item{refName}{group name for reference samples}

\item{metadata}{an optional data frame with metadata items as columns, where samples are in the same order as in abundances and data types (factor vs numeric) are supposed to be correct; if provided and rda is FALSE, envfit is carried out}

\item{groupAttrib}{optional: the name of a metadata item that refers to a vector that provides for each sample its group membership}

\item{groups}{an optional vector that provides for each sample its group membership and which is overridden by groupAttrib, if provided}

\item{groupColors}{an optional map of predefined colors for groups that matches names in groups (which should be strings); if reference is provided, refName is added if absent}

\item{colors}{an optional vector of colors to be used to color samples; it overrides groupColors if provided}

\item{clusters}{an optional vector that provides for each sample its cluster membership (cluster membership is visualized through shape, up to 10 different shapes are possible)}

\item{labels}{an optional vector that provides for each sample a label to display}

\item{sizes}{a vector of  numeric values that will be displayed as varying sample sizes (sizes will be shifted into positive range if necessary and scaled between 0.5 and 2.5)}

\item{size.legend}{a string displayed as a legend for size}

\item{time}{an optional vector with as many time points as samples, adds arrows between consecutive time points (time points should be in ascending order)}

\item{hiddenSamples}{an optional vector with indices of samples to be hidden (they will be taken into account for PCoA/RDA/envfit, but are not displayed)}

\item{dis}{dissimilarity or distance supported by vegan's vegdist function (if set to cor, a PCA is carried out using vegan's function rda with scale set to true)}

\item{rda}{carry out an RDA instead of a PCoA using vegan's capscale function}

\item{scale}{scale numeric metadata (subtract the mean and divide by standard deviation)}

\item{doScree}{do a Scree plot}

\item{topTaxa}{if larger than zero: show the top N taxa most strongly covarying with principal components as arrows in the PCoA if they are significant according to a permutation test}

\item{topMetadata}{if larger than zero, metadata provided and rda false: show the top N most significant numeric metadata as arrows and the top N most significant factor metadata as text in the PCoA}

\item{arrowFactor}{the length of taxon arrows (determined by scaled covariance) is multiplied with this factor}

\item{metadataFactor}{the length of numeric metadata arrows (determined by Pearson correlation) is multiplied with this factor}

\item{centroidFactor}{centroid positions (representing categoric metadata) are multiplied with this factor}

\item{taxonColor}{the color of the taxon arrows and text}

\item{metadataColor}{the color of the metadata arrows and text}

\item{xlim}{range shown on the x axis, by default the minimum and maximum of the first selected component}

\item{ylim}{range shown on the y axis, by default the minimum and maximum of the second selected component}

\item{permut}{number of permutations for top-covarying taxa; if NA, NULL or smaller than 1, no permutation test is carried out}

\item{env.permut}{number of permutations for envfit}

\item{pAdjMethod}{method for multiple testing correction supported by p.adjust for top-covarying taxon and envfit p-values}

\item{qvalThreshold}{threshold on multiple-testing corrected top-covarying taxon and envfit p-values}

\item{dimensions}{the principal components used for plotting, by default the first and second}

\item{\\dots}{Additional arguments passed to plot()}
}
\description{
A wrapper around various PCoA-based analyses implemented in vegan. The wrapper can handle groups and
metadata. PCoA is carried out sample-wise. The na.action is set to na.omit, however envfit cannot deal with
missing values, therefore if metadata are provided, they should be free of missing values.
}
\details{
When a reference and groups are provided and the number of group memberships does not equal the number of samples in the combined abundance table,
groups are automatically extended such that reference samples are assigned to a single group with name refName, which is colored in gray.
The color vector is likewise extended if provided. If a clusters, time and/or labels vector is provided together with a referene, it has to refer to both data sets.
Samples in the abundance matrix are appended after the reference samples, so cluster memberships, time points and/or labels have to be provided in the same order.
Different total counts in abundances and reference samples may bias the result, therefore rarefyRef allows to rarefy both to the same total count after matching.
RarefyRef will rarefy all samples to the lowest total count found in any sample. Rows with zero counts after rarefaction are removed.
When topTaxa is set larger zero, significant top-varying taxa are shown. The permutation test is carried out by shuffling the selected number of top-covarying taxa.
Multiple testing correction on parameter-free p-values is then only applied to these top-covarying taxa. The strength of covariance is determined as
the norm of the vectors resulting from multiplying the standardized eigen vectors with the taxa.
In contrast, envfit p-values are computed for all metadata and multiple-testing correction is consequently applied to all metadata provided, though
only the selected number of most significant metadata are shown. Thus, topTaxa ranks taxa by covariance with significance as a filter, whereas
topMetadata ranks metadata by significance.
The number in the axis label brackets refers to the proportion of variance explained as computed with vegan's eigenvals function.
}
\examples{
data("ibd_taxa")
data("ibd_metadata")
ibd_metadata=assignMetadataTypes(ibd_metadata,categoric=c("SRA_metagenome_name","Diagnosis"))
seqPCoA(ibd_taxa,groups=as.vector(ibd_metadata$Diagnosis),topTaxa=30)
# remove 65 samples with missing calprotectin measurements or other missing values in the metadata
na.indices=unique(which(is.na(ibd_metadata),arr.ind=TRUE)[,1])
indices.to.keep=setdiff(1:nrow(ibd_metadata),na.indices)
ibd_metadata=ibd_metadata[indices.to.keep,]
ibd_taxa=ibd_taxa[,indices.to.keep]
seqPCoA(ibd_taxa,metadata=ibd_metadata,groups=as.vector(ibd_metadata$Diagnosis),topTaxa=30)
}
