% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{Get silhouette score for cluster labels}
\usage{
silhouette(abundances, groups, method = "bray")
}
\arguments{
\item{abundances}{a matrix with taxa as rows and samples as columns}

\item{groups}{vector with group memberships}

\item{method}{name of dissimilarity or distance to use. See vegdist for options
#' @export}
}
\description{
A wrapper around vegdist that computes the silhouette score directly on these distances
}
